Add a bunch of variant "only" information to varcon.  That is information on variants with no non-American entry.  For example:   _: whimsy / _V: whimsey These new entries are flagged with "_" since they likely apply to British English, but the information was derived from American sources.
Adjust varies scripts is varcon and scowl to be able to work with the
new data.  In the process I tried to make the variant_* spelling
category more useful for non-American spellings.  However, what is
really needed is separate variants for each of the spelling categories
(ie american-v0, etc).  Who knows when this will get done.  It
probably isn't much work but it will likely increase the number of
lists by up to a factor of 4, and will also effect how the Aspell
dictionaries are created.  This might wait until SCOWL is rewritten to
work with something other than lots and lots of, mostly unstructured,
word lists.

Also add most of the raw data and scripts I used to create the new
entries under varcon/raw/.

This work was finished primary back in March.  I think the motivation
behind it was that there where a lot of variants in alt12dicts which I
did not have entries for in varcon.txt or variant-also.tab, which
created a bit of a problem.  Rather than simply add some rows to
variant-also.tab I decided to take the time too "do the correct
thing". ... In the process I also made a lot of entries in
variant-also.tab redundant, which this commit also removes.

Some other unrelated changes in scowl and likely varcon, which where
made sometime in the past year.

[svn rev 99]