Use symbolic links to avoid duplicating readme text when possible.
Also remove the "text-file" layout and replace it with a "readme-text"
layout.  This new layout will automatically include the text of the
readme file in the body.

For now leave the readme for pos, jargon-wl, and ispell-enwl unwrapped.
These projects have not been updated in a long time and contain outdated
links back to the wordlist (i.e. SCOWL and friends) project.