Optimize and enhance "speller_lookup.pm".
Optimize module so it would scale better for large lists.

Enhance module to handle the case general when another capitalization
of a word is in the dictionary; for example, looking up "york" when
"York" is the dictionary.

Also do not include SCOWL size 95 in the results as that size has no
quality control.