Add a list of not so common contractions, but don't include them just yet.
Alan wrote on issue #127: The contractions are all common in speech.
I must admit that mostly I've never seen these in writing, however.
How'd and how're are the only ones that seem remotely common in writing.
(But I don't think it would be a bad thing to add them all.)