merge-groups: various enhancements
Add ability to remove lemmas (and all derived words) from a group.

Add ability to add lemmas with derived forms (as oppose to just the lemma).

Add ability to adjust group info.  This included changing the pos and
defn_note.  When changing the pos derived entries will automatically get mapped
to the corrected pos when possibe.