libscowl: refactor
Move tags into there own class for additional functionality.

Factor out most of the functionality of exportDB into _exportGroup so that it
is possible to export a single group at a time.

Add merge method to Groups and use it instead of an inner function of the
same name inside the LineBase.parse method.